SpectralLearning

Documentation for SpectralLearning.

SpectralLearning.AbstractOperator

SpectralLearning.FourierKernel

SpectralLearning.InnerProduct

SpectralLearning.KernelOperator

SpectralLearning.PolynomialKernel

SpectralLearning.Spectral

SpectralLearning.spectralize

SpectralLearning.AbstractOperator — Type

AbstractOperator

An abstract type for linear operators on Banach spaces.

SpectralLearning.FourierKernel — Type

FourierKernel{T<:Real}

An object representing a Fourier kernel function $k(x,y) = sin (ω π xy + θ)$

SpectralLearning.InnerProduct — Type

InnerProduct{F,S}

An struct representing a linear form on a Hilbert space as an inner product $⟨w, x⟩$.

SpectralLearning.KernelOperator — Type

KernelOperator{K,S} <: AbstractOperator

A type for representing kernel integral operators over Hilbert spaces.

SpectralLearning.PolynomialKernel — Type

PolynomialKernel{M<:AbstractMatrix}

A type representing a polynomial kernel function.

SpectralLearning.Spectral — Type

Spectral{F,M,V1,V2,V3}

Create a spectral layer, whose forward pass is given by

σ.(B*(λ_in.*x) .- λ_out.*(B*x) .+ b)

where B is a static matrix, and λ_in, λ_out and b are trainable vectors of parameters.

These layers can be used as drop-in replacement to traditional dense layers, and are usually as good as dense layers for similar amounts of parameters (but spectral layers are wider for the same number of parameters).

Examples

julia> sp = Spectral(5, 2)
Spectral(5 => 2)

SpectralLearning.spectralize — Method

spectralize(d::Dense)

Create a spectral layer with the same dimensions as d.